requirements.txt
openai 
fastapi 
sse-starlette 
uvicorn
python-dotenv
nltk
tiktoken
langchain_openai
langchain